requirements.txt
streamlit
pdfminer.six
ebooklib
requests
